+++ services/network/mgmt/2020-11-01/network/CHANGELOG.md
@@ -1,2 +1,30 @@
-# Change History
+# Unreleased
 
+## Breaking Changes
+
+### Struct Changes
+
+#### Removed Struct Fields
+
+1. BackendAddressPoolPropertiesFormat.Location
+
+## Additive Changes
+
+### New Constants
+
+1. BastionHostSkuName.BastionHostSkuNameBasic
+1. BastionHostSkuName.BastionHostSkuNameStandard
+
+### New Funcs
+
+1. PossibleBastionHostSkuNameValues() []BastionHostSkuName
+
+### Struct Changes
+
+#### New Structs
+
+1. Sku
+
+#### New Struct Fields
+
+1. BastionHost.Sku